requirements.txt
PyPDF2
pdf2image
azure-storage-blob
python-dotenv
openai==1.1.2